Moved the components from the news route to the generic components folder as they can be reused.
