+++ app/components/EventSummary.tsx
