+++ app/components/ProductSummary.translations.ts
