+++ app/components/ProductSummary.tsx
