+++ app/components/ProjectSummary.translations.ts
