+++ app/components/ProjectSummary.tsx
