+++ app/components/YearSelector.translations.ts
